pyproject.toml
[tool.poetry]
name = "syneto_labs"
version = "0.1.0"
description = "Python Time Machine"
authors = ["Alex Serban"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
click = "^7.1.2"
yfinance = "^0.1.59"
matplotlib = "^3.4.1"
aiofiles = "^0.6.0"
yagmail = "^0.14.256"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
